Testing

+++ src/main/java/com/adp/domain/JobTransferRequest.java
@@ -3,7 +3,7 @@ import lombok.Data;
 
 @Data
 public class JobTransferRequest {
-    public int jobId;
-    public int fromUserId;
-    public int toUserId;
+    public Long jobId;
+    public Long fromUserId;
+    public Long toUserId;
 }
